fix: function in filter
